make all filenames use only lower-case letters
