+++ example/advection/advection.cpp
@@ -23,8 +23,8 @@
 #include "bvals/boundary_conditions.hpp"
 #include "bvals/bvals.hpp"
 #include "driver/multistage.hpp"
-#include "interface/Params.hpp"
-#include "interface/StateDescriptor.hpp"
+#include "interface/params.hpp"
+#include "interface/state_descriptor.hpp"
 #include "mesh/mesh.hpp"
 #include "parthenon_manager.hpp"
 #include "reconstruct/reconstruction.hpp"
