+++ example/advection/advection.hpp
@@ -17,8 +17,8 @@
 
 #include "driver/driver.hpp"
 #include "driver/multistage.hpp"
-#include "interface/Container.hpp"
-#include "interface/StateDescriptor.hpp"
+#include "interface/container.hpp"
+#include "interface/state_descriptor.hpp"
 #include "mesh/mesh.hpp"
 #include "task_list/tasks.hpp"
