+++ example/face_fields/face_fields_example.hpp
@@ -18,7 +18,7 @@
 
 #include "driver/driver.hpp"
 #include "globals.hpp"
-#include "interface/StateDescriptor.hpp"
+#include "interface/state_descriptor.hpp"
 #include "mesh/mesh.hpp"
 #include "task_list/tasks.hpp"
