+++ src/bvals/boundary_conditions.cpp
@@ -14,8 +14,8 @@
 #include "bvals/boundary_conditions.hpp"
 
 #include "bvals/bvals_interfaces.hpp"
-#include "interface/Container.hpp"
-#include "interface/ContainerIterator.hpp"
+#include "interface/container.hpp"
+#include "interface/container_iterator.hpp"
 #include "mesh/mesh.hpp"
 
 namespace parthenon {
