+++ src/interface/metadata.cpp
@@ -11,7 +11,7 @@
 // the public, perform publicly and display publicly, and to permit others to do so.
 //========================================================================================
 
-#include "interface/Metadata.hpp"
+#include "interface/metadata.hpp"
 
 #include <exception>
 #include <sstream>
