+++ src/interface/sparse_variable.cpp
@@ -10,11 +10,32 @@
 // license in this material to reproduce, prepare derivative works, distribute copies to
 // the public, perform publicly and display publicly, and to permit others to do so.
 //========================================================================================
-#include "interface/PropertiesInterface.hpp"
+
+#include "interface/sparse_variable.hpp"
+
+#include "interface/metadata.hpp"
 
 namespace parthenon {
 
-// Initialize the static map of ids
-std::map<std::string, int> PropertiesInterface::_label_to_id;
+template <typename T>
+void SparseVariable<T>::Add(int varIndex) {
+  // Now allocate depending on topology
+  if ((metadata_.Where() == Metadata::Cell) || (metadata_.Where() == Metadata::Node)) {
+    // check if variable index already exists
+    if (varMap_.find(varIndex) != varMap_.end()) {
+      throw std::invalid_argument("Duplicate index in create SparseVariable");
+    }
+    // create the variable and add to map
+    std::string my_name = label_ + "_" + std::to_string(varIndex);
+    auto v = std::make_shared<CellVariable<T>>(my_name, dims_, metadata_);
+    varArray_.push_back(v);
+    indexMap_.push_back(varIndex);
+    varMap_[varIndex] = v;
+  } else {
+    throw std::invalid_argument("unsupported type in SparseVariable");
+  }
+}
+
+template class SparseVariable<Real>;
 
 } // namespace parthenon
