+++ src/interface/sparse_variable.hpp
@@ -15,8 +15,8 @@
 /// Builds on ParArrayNDs
 /// Date: Sep 12, 2019
 ///
-#ifndef INTERFACE_SPARSEVARIABLE_HPP_
-#define INTERFACE_SPARSEVARIABLE_HPP_
+#ifndef INTERFACE_SPARSE_VARIABLE_HPP_
+#define INTERFACE_SPARSE_VARIABLE_HPP_
 
 #include <functional>
 #include <iostream>
@@ -27,7 +27,7 @@
 #include <vector>
 
 #include "globals.hpp"
-#include "interface/Variable.hpp"
+#include "interface/variable.hpp"
 
 namespace parthenon {
 
@@ -140,4 +140,4 @@ using MapToSparse = std::map<std::string, std::shared_ptr<SparseVariable<T>>>;
 
 } // namespace parthenon
 
-#endif // INTERFACE_SPARSEVARIABLE_HPP_
+#endif // INTERFACE_SPARSE_VARIABLE_HPP_
