+++ src/interface/update.cpp
@@ -11,14 +11,14 @@
 // the public, perform publicly and display publicly, and to permit others to do so.
 //========================================================================================
 
-#include "interface/Update.hpp"
+#include "interface/update.hpp"
 
 #include <algorithm>
 #include <limits>
 
 #include "coordinates/coordinates.hpp"
-#include "interface/Container.hpp"
-#include "interface/ContainerIterator.hpp"
+#include "interface/container.hpp"
+#include "interface/container_iterator.hpp"
 #include "mesh/mesh.hpp"
 
 namespace parthenon {
