+++ src/interface/update.hpp
@@ -14,7 +14,7 @@
 #define INTERFACE_UPDATE_HPP_
 
 #include "athena.hpp"
-#include "interface/Container.hpp"
+#include "interface/container.hpp"
 #include "mesh/mesh.hpp"
 
 namespace parthenon {
