+++ src/interface/variable.hpp
@@ -36,7 +36,7 @@
 #include "athena.hpp"
 #include "basic_types.hpp"
 #include "bvals/cc/bvals_cc.hpp"
-#include "interface/Metadata.hpp"
+#include "interface/metadata.hpp"
 #include "parthenon_arrays.hpp"
 
 namespace parthenon {
