+++ src/mesh/mesh.hpp
@@ -31,11 +31,11 @@
 #include "athena.hpp"
 #include "bvals/bvals.hpp"
 #include "bvals/bvals_interfaces.hpp"
-#include "interface/Container.hpp"
-#include "interface/PropertiesInterface.hpp"
-#include "interface/StateDescriptor.hpp"
-#include "interface/Update.hpp"
+#include "interface/container.hpp"
 #include "interface/container_collection.hpp"
+#include "interface/properties_interface.hpp"
+#include "interface/state_descriptor.hpp"
+#include "interface/update.hpp"
 #include "kokkos_abstraction.hpp"
 #include "mesh/mesh_refinement.hpp"
 #include "mesh/meshblock_tree.hpp"
