+++ src/mesh/meshblock.cpp
@@ -32,9 +32,9 @@
 #include "bvals/bvals.hpp"
 #include "coordinates/coordinates.hpp"
 #include "globals.hpp"
-#include "interface/ContainerIterator.hpp"
-#include "interface/Metadata.hpp"
-#include "interface/Variable.hpp"
+#include "interface/container_iterator.hpp"
+#include "interface/metadata.hpp"
+#include "interface/variable.hpp"
 #include "kokkos_abstraction.hpp"
 #include "mesh/mesh.hpp"
 #include "mesh/mesh_refinement.hpp"
