+++ src/outputs/athena_hdf5_C.cpp
@@ -25,7 +25,7 @@
 #include "athena.hpp"
 #include "coordinates/coordinates.hpp"
 #include "globals.hpp"
-#include "interface/ContainerIterator.hpp"
+#include "interface/container_iterator.hpp"
 #include "mesh/mesh.hpp"
 #include "outputs/outputs.hpp"
 #include "parameter_input.hpp"
